face-api/requirements.txt
fastapi
uvicorn[standard]
face_recognition
numpy
Pillow
